Evo_requirements.txt.txt
stripedhyena==0.2.2
biopython
pandas
accelerate
datasets
trl
peft